feat(iam): move challenge into its own file
