+++ iam/src/utils/providers.ts
@@ -1,9 +1,6 @@
 // ---- Types
 import type { Provider } from "../types";
-import type { RequestPayload, ChallengePayload, VerifiedPayload, ProviderContext } from "@gitcoin/passport-types";
-
-// ---- Return randomBytes as a challenge to test that the user has control of a provided address
-import crypto from "crypto";
+import type { RequestPayload, VerifiedPayload, ProviderContext } from "@gitcoin/passport-types";
 
 // Collate all Providers to abstract verify logic
 export class Providers {
@@ -20,41 +17,6 @@ export class Providers {
 
       return providers;
     }, {} as { [k: string]: Provider });
-  }
-
-  // request a challenge sig
-  getChallenge(payload: RequestPayload): ChallengePayload {
-    // @TODO - expand this to allow providers to set custom challanges?
-    const getChallengeString = (provider: string, nonce: string): string => {
-      switch (provider) {
-        case "SignerChallenge":
-          return `I commit that this wallet is under my control and that I wish to link it with my Passport.\n\nnonce: ${nonce}`;
-          break;
-        case "Signer":
-          return `I commit that I wish to register all ETH stamps associated with an Ethereum account that I control to my Passport.\n\naccount: ${payload.signer.address}\nnonce: ${nonce}`;
-        default:
-          return `I commit that this stamp is my unique and only ${provider} verification for Passport.\n\nnonce: ${nonce}`;
-          break;
-      }
-    };
-    // check that we've been provided an address for the challenge
-    if (payload.address) {
-      // valid payload - create a challenge string
-      return {
-        valid: true,
-        record: {
-          address: payload.address,
-          type: payload.type,
-          challenge: getChallengeString(payload.type, crypto.randomBytes(32).toString("hex")),
-        },
-      };
-    } else {
-      // unable to create a challenge without address
-      return {
-        valid: false,
-        error: ["Missing address"],
-      };
-    }
   }
 
   // Given the payload is valid return the response of the selected Providers verification proceedure
